+++ src/utils/background/getExtensionIconPath.ts
