+++ src/utils/background/getPreferredColorScheme.ts
