+++ src/utils/background/getSpecialIconPath.ts
