+++ src/utils/background/index.ts
@@ -1,0 +1,21 @@
+export * from "./onUrlChange";
+export * from "./removeBadge";
+export * from "./showExtensionIcon";
+export * from "./getSpecialIconPath";
+export * from "./showBadgeIfAllowed";
+export * from "./setIconsForAllSizes";
+export * from "./showCountryFlagIcon";
+export * from "./getExtensionIconPath";
+export * from "./showOnionAddressIcon";
+export * from "./getPreferredColorScheme";
+
+export { default as onUrlChange } from "./onUrlChange";
+export { default as removeBadge } from "./removeBadge";
+export { default as showExtensionIcon } from "./showExtensionIcon";
+export { default as getSpecialIconPath } from "./getSpecialIconPath";
+export { default as showBadgeIfAllowed } from "./showBadgeIfAllowed";
+export { default as setIconsForAllSizes } from "./setIconsForAllSizes";
+export { default as showCountryFlagIcon } from "./showCountryFlagIcon";
+export { default as getExtensionIconPath } from "./getExtensionIconPath";
+export { default as showOnionAddressIcon } from "./showOnionAddressIcon";
+export { default as getPreferredColorScheme } from "./getPreferredColorScheme";
