+++ src/utils/background/removeBadge.ts
