+++ src/utils/index.ts
@@ -1,22 +1,24 @@
-export * from "./buildUrl";
+export * from "./isOnionAddress";
+export * from "./isLocalHostAddress";
+export * from "./getCurrentTab";
+export * from "./domainData";
+export * from "./options";
+export * from "./translate";
+export * from "./getStorageMethod";
 export * from "./getFlagPath";
-export * from "./getPreferredColorScheme";
+export * from "./buildUrl";
+export * from "./instance";
+export * from "./getBrowserLanguageCode";
 
 export { default as buildUrl } from "./buildUrl";
-export { default as getCurrentTab } from "./getCurrentTab";
-export { default as getFlagPath } from "./getFlagPath";
-export { default as getPreferredColorScheme } from "./getPreferredColorScheme";
-export { default as getSpecialIconPath } from "./getSpecialIconPath";
-export { default as isOnionAddress } from "./isOnionAddress";
-export { default as isLocalHostAddress } from "./isLocalHostAddress";
-export { default as onUrlChange } from "./onUrlChange";
 export { default as instance } from "./instance";
 export { default as getBrowserLanguageCode } from "./getBrowserLanguageCode";
-export { default as getExtensionIconPath } from "./getExtensionIconPath";
+export { default as getCurrentTab } from "./getCurrentTab";
+export { default as isOnionAddress } from "./isOnionAddress";
+export { default as isLocalHostAddress } from "./isLocalHostAddress";
 export { default as translate } from "./translate";
-export { default as showBadgeIfAllowed } from "./showBadgeIfAllowed";
-export { default as removeBadge } from "./removeBadge";
-export { default as setIconsForAllSizes } from "./setIconsForAllSizes";
 export { default as getStorageMethod } from "./getStorageMethod";
+export { default as getFlagPath } from "./getFlagPath";
+
 export * as options from "./options";
 export * as domainData from "./domainData";
